+++ src/KeyVault/KeyVault/SecurityDomain/Cmdlets/ImportAzKeyVaultSecurityDomain.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.Azure.Commands.Common.Authentication;
+using Microsoft.Azure.Commands.Common.Authentication;
 using Microsoft.Azure.Commands.Common.Exceptions;
 using Microsoft.Azure.Commands.KeyVault.Models;
 using Microsoft.Azure.Commands.KeyVault.Properties;