+++ src/Accounts/Accounts/ChangeLog.md
@@ -19,7 +19,8 @@
 -->
 
 ## Upcoming Release
-* Fixed an issue that the usage of some cmdlets are not counted.
+* Added the module name in breaking change messages 
+* Upgraded Microsoft.ApplicationInsights version from 2.13.1 to 2.18.0 
 
 ## Version 2.13.0
 * Supported in-tool notification for version upgrade.